+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- `@Bound` annotation (works like `@BoundTo`, but implicitly uses the direct supertype of the annotated class)
+
 ### Changed
 - Make `component` in annotations optional and set to `SingletonComponent` by default.
 - Dependencies: